Move `create-dummy-tables` tests
